perf(note): graph weight use float32 not float 64

+++ internal/note/app/model.go
@@ -68,7 +68,7 @@ type GraphNode struct {
 	ID     string
 	Name   string
 	Type   GraphNodeType
-	Weight float64
+	Weight float32
 }
 
 type GraphLink struct {
